requirements.txt
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
jupyter>=1.0.0
scikit-learn>=1.2.0
tqdm>=4.65.0
pandas>=2.0.0
seaborn>=0.12.0
Pillow>=10.0.0
pytest>=7.4.0
